AutoConnectEvent, StaticTuple
The 'auto connect event' tool simplifies connecting event publishers
with event subscribers by matching events with correctly named and
argumented private methods.

The 'static tuple' tool is similar to a standard tuple, except that it
provides Equals functionality that only checkes for reference equality,
good for tuple of type used as keys in dictionary for example.